test: move the root-resolution reader tests into their own file
reader_test.go covers transcript parsing; these four cover which roots the
reader searches. Both concerns were appending to the same file's tail, which
made this branch collide with unrelated parsing work on that tail for no reason
other than proximity - the code under test does not overlap at all.

The helper hardening stays in reader_test.go, where the helper lives: clearing
the HERDR_*_CONFIG_DIRS variables protects every reader test from a developer's
exported environment, not only the four that moved.
